+++ out/extension.js
@@ -34,7 +34,7 @@ let config = vscode.workspace.getConfiguration('stardew-pets');
 let webview;
 let extensionStorageFolder = '';
 //Enums
-const EnemySpecies = {
+const MonsterSpecies = {
     Slime: ['Iron', 'Tiger'],
     Bug: ['Normal', 'Normal Dangerous', 'Armored', 'Armored Dangerous'],
     Golem: ['Stone', 'Stone Dangerous', 'Iridium', 'Wilderness'],
@@ -142,6 +142,11 @@ function initGame() {
     webview.postMessage({
         type: 'scale',
         value: config.get('scale')
+    });
+    //Send monsters toggle
+    webview.postMessage({
+        type: 'monsters',
+        value: config.get('monsters')
     });
     //Send money
     webview.postMessage({
@@ -285,6 +290,13 @@ function activate(context) {
             webview.postMessage({
                 type: 'scale',
                 value: config.get('scale')
+            });
+        }
+        //Monsters toggle changed
+        if (event.affectsConfiguration("stardew-pets.monsters")) {
+            webview.postMessage({
+                type: 'monsters',
+                value: config.get('monsters')
             });
         }
     });
@@ -454,17 +466,17 @@ class WebViewProvider {
                     save.money = message.value;
                     saveGame();
                     break;
-                //Spawn enemy
-                case 'spawn_enemy': {
+                //Spawn monster
+                case 'spawn_monster': {
                     //Get specie
-                    const species = Object.keys(EnemySpecies);
+                    const species = Object.keys(MonsterSpecies);
                     const specie = species[Math.floor(Math.random() * species.length)];
                     //Get color
-                    const colors = EnemySpecies[specie];
+                    const colors = MonsterSpecies[specie];
                     const color = colors[Math.floor(Math.random() * colors.length)];
-                    //Spawn enemy
+                    //Spawn monster
                     this.postMessage({
-                        type: 'spawn_enemy',
+                        type: 'spawn_monster',
                         specie: specie,
                         color: color,
                     });